cluster_get_chain_list: Now prints out notes for first task cluster_print: changed mem to type double cluster_stop: corrected location of ctrl.cluster.jm check tracker_threshold: better input checks, 2D multilooking update_surface_with_tracker: better input checks, added trim feature make_gps_2018_greenland_P3.m: setup all gps vars default_radar_params_2018_Antarctica_TObas_accum3: superficial updates publish_echogram: Support ~zero fast-time length echograms analysis*: adjusted mem requirements, fixed header comments create_records*: preliminary support for snow radar run_collate_deconv arena_packet_strip_task: changed rx_paths to be per wf variable preprocess: new standard for preprocessing (replaces create_segments*, run_browse_ni*) Updated field default radar params

+++ cresis-toolbox/cluster/cluster_print.m
@@ -465,13 +465,13 @@ if print_mode == 2
             tmp_result = tmp_result(idx:end);
             mem_units = sscanf(tmp_result,'%s');
             if strcmpi(mem_units,'mb')
-              info(id_idx).mem_actual = uint32(mem);
+              info(id_idx).mem_actual = mem;
             elseif strcmpi(mem_units,'kb')
-              info(id_idx).mem_actual = uint32(mem/1e3);
+              info(id_idx).mem_actual = mem/1e3;
             elseif strcmpi(mem_units,'gb')
-              info(id_idx).mem_actual = uint32(mem*1e3);
+              info(id_idx).mem_actual = mem*1e3;
             else
-              info(id_idx).mem_actual = uint32(mem/1e6);
+              info(id_idx).mem_actual = mem/1e6;
             end
           end
           
@@ -634,7 +634,7 @@ if print_mode == 2
           idx = find(tmp_result==10|tmp_result==13,1);
           tmp_result = tmp_result(1:idx);
           [mem,~,~,idx] = sscanf(tmp_result,'%d');
-          info(id_idx).mem_actual = uint32(mem/1e3);
+          info(id_idx).mem_actual = mem/1e3;
         end
       end
     end
